+++ assets/javascripts/kitten/components/cards/article-card.js
@@ -100,15 +100,22 @@ class ArticleCardComponent extends Component {
   }
 
   renderTitle() {
+    const className = classNames(
+      'k-Card__title',
+      this.props.titleProps.className,
+    )
+
     return (
       <Marger
         bottom="1"
         style={ styles.title }
       >
         <Title
+          tag="p"
+          { ...this.props.titleProps }
           modifier="senary"
           margin={ false }
-          className="k-Card__title"
+          className={ className }
         >
           { parseHtml(this.props.articleTitle) }
         </Title>
@@ -191,6 +198,7 @@ ArticleCardComponent.defaultProps = {
     src: 'https://placehold.it/100x100/4a84ff/4a84ff',
     alt: '',
   },
+  titleProps: {},
   ownerTitle: 'Owner title',
   ownerDescription: 'Owner description',
   articleTitle: 'Title',