+++ assets/javascripts/kitten/components/cards/simple-card.test.js
@@ -125,4 +125,26 @@ describe('<SimpleCard />', () => {
       })
     })
   })
+
+  describe('with title props', () => {
+    beforeEach(() => {
+      component = mount(
+        <SimpleCard
+          title="Simple card title"
+          titleProps={{
+            tag: 'h2',
+            className: 'custom-class',
+          }}
+        />
+      )
+    })
+
+    it('has a title with h2 tag', () => {
+      expect(component.find(Title)).to.have.prop('tag', 'h2')
+    })
+
+    it('has a custom class on Title', () => {
+      expect(component.find(Title)).to.have.className('custom-class')
+    })
+  })
 })